requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
scrapy==2.11.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
schedule==1.2.0
geopy==2.4.1
newspaper3k==0.2.8
feedparser==6.0.10
python-dateutil==2.8.2
pytz==2023.3
celery==5.3.4
redis==5.0.1
flower==2.0.1
aiohttp==3.9.1

